feat: Join Api dev-1

+++ src/main/java/com/news/dev/api/user/entity/UserEntity.java
@@ -12,7 +12,7 @@ import javax.persistence.*;
 public class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userNo;
+    private Long userNo;
 
     @Column(length = 250, unique = true)
     private String email;
